examples/medqa_multimodel/medqa_multimodel_requirements.txt
openai>=1.0.0
anthropic>=0.49.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
benchflow
